created hook for disconnect so we can use it wherever required.
